Initial Python3 support
