+++ test/functional/test_pipemode.py
@@ -38,7 +38,7 @@ def recordio_file():
 def tfrecords_file():
     writer = tf.python_io.TFRecordWriter("test.tfrecords")
     for i in range(100):
-        writer.write("hello world")
+        writer.write(b"hello world")
     writer.close()
     yield
     os.remove('test.tfrecords')
@@ -157,7 +157,7 @@ def test_tf_record():
         it = ds.make_one_shot_iterator()
         next = it.get_next()
         for i in range(100):
-            assert sess.run(next) == 'hello world'
+            assert sess.run(next) == b'hello world'
 
 
 FIELD_DEFAULTS = [[0] for i in range(100)]
