+++ test/integ/Dockerfile-py3
@@ -1,0 +1,59 @@
+FROM ubuntu:16.04
+
+RUN apt-get update && apt-get install -y --no-install-recommends \
+        build-essential \
+        curl \
+        git \
+        libcurl3-dev \
+        libfreetype6-dev \
+        libpng12-dev \
+        libzmq3-dev \
+        pkg-config \
+        rsync \
+        software-properties-common \
+        unzip \
+        zip \
+        zlib1g-dev \
+        openjdk-8-jdk \
+        openjdk-8-jre-headless \
+        wget \
+        vim \
+        iputils-ping \
+        nginx \
+        && \
+    apt-get clean && \
+    rm -rf /var/lib/apt/lists/*
+
+RUN apt-get install -y python3
+
+RUN curl -fSsL -O https://bootstrap.pypa.io/get-pip.py && \
+    python3 get-pip.py && \
+    rm get-pip.py
+
+RUN python3 -m pip install cmake
+
+ARG device=cpu
+ARG tensorflow_version=1.9.0
+ARG script
+
+RUN if [ "$device"="cpu" ] ; then \
+		python3 -m pip install tensorflow==$tensorflow_version; \
+	else \
+		python3 -m pip install "tensorflow-gpu"==$tensorflow_version; \
+	fi
+
+ARG sagemaker_tensorflow=sagemaker_tensorflow-$tensorflow_version.1.0.0.tar.gz
+
+WORKDIR /root
+
+COPY $sagemaker_tensorflow .
+
+RUN sagemaker_tensorflow_local=$(basename $sagemaker_tensorflow) && \
+    \
+    python3 -m pip install $sagemaker_tensorflow_local && \
+    \
+    rm $sagemaker_tensorflow_local
+
+COPY $script script.py
+
+ENTRYPOINT ["python3", "script.py"]
