+++ tox.ini
@@ -1,5 +1,6 @@
 [tox]
-envlist = py27,flake8
+envlist = py27,py34,py35,py36,flake8
+
 skip_missing_interpreters = False
 
 [flake8]
